fix assets & shims
