+++ packages/base/package-scripts.js
@@ -10,7 +10,7 @@ const scripts = {
 		bundle: "rollup --config config/rollup.config.js --environment ES5_BUILD",
 	},
 	copy: {
-		default: "nps copy.cldr copy.src copy.test copy.webcomponents-polyfill",
+		default: "nps copy.src copy.test copy.webcomponents-polyfill",
 		cldr: 'copy-and-watch "../../node_modules/@ui5/webcomponents-utils/dist/**/cldr/*.json" dist/generated/assets/cldr/',
 		src: "copy-and-watch \"src/**/*.js\" dist/",
 		test: "copy-and-watch \"test/**/*.*\" dist/test-resources",
