+++ packages/base/src/asset-registries/i18n.js
@@ -1,4 +1,3 @@
-import "../Shims.js";
 import getLocale from "../locale/getLocale.js";
 import { fetchJsonOnce } from "../util/FetchHelper.js";
 import normalizeLocale from "../locale/normalizeLocale.js";
