+++ packages/utils/src/DateFormat.js
@@ -1,0 +1,4 @@
+import "./Shims.js";
+import DateFormat from "./sap/ui/core/format/DateFormat.js";
+
+export default DateFormat;
