+++ packages/utils/src/LocaleData.js
@@ -1,0 +1,4 @@
+import "./Shims.js";
+import LocaleData from "./sap/ui/core/LocaleData.js";
+
+export default LocaleData;
