+++ packages/utils/src/features/calendar/Gregorian.js
@@ -1,0 +1,1 @@
+import "../../sap/ui/core/date/Gregorian.js";
